Move document assertions into trait
